+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.2.7
+  * Checks Access Token expiration before attempting to refresh token.
+  * Adds backoff case for token expiration.
+  * Adds unittests and unittest related functions.
+  [#50](https://github.com/singer-io/tap-linkedin-ads/pull/50)
+
 ## 1.2.6
   * Fixed 1.2.5 release issues [#43](https://github.com/singer-io/tap-linkedin-ads/pull/43)